refactor(dto): Refactor modules dtos
