+++ src/modules/assunto/dto/updateAssunto.dto.ts
@@ -1,5 +1,5 @@
 import { ApiModelPropertyOptional, ApiModelProperty } from '@nestjs/swagger';
-import { IsInt, IsDefined, IsString, IsArray } from 'class-validator';
+import { IsInt, IsDefined, IsString, IsArray, IsOptional } from 'class-validator';
 import { UpdateFonteDto } from '../../fonte/dto/index';
 import { UpdateDescritorDto } from '../../descritor/dto/index';
 
@@ -14,12 +14,14 @@ export class UpdateAssuntoDto {
   @ApiModelPropertyOptional({
     description: 'Nome do assunto',
   })
+  @IsOptional()
   @IsString()
   nome: string;
 
   @ApiModelPropertyOptional({
     description: 'Descrição do assunto',
   })
+  @IsOptional()
   @IsString()
   descricao: string;
 }
