+++ src/modules/fonte/dto/createFonte.dto.ts
@@ -1,5 +1,5 @@
 import { ApiModelProperty, ApiModelPropertyOptional } from '@nestjs/swagger';
-import { IsString } from 'class-validator';
+import { IsString, IsOptional } from 'class-validator';
 import { TipoFonteDto } from '../../tipoFonte/dto/index';
 import { Assunto } from '../../../database/entities/assunto.entity';
 import { Descritor } from '../../../database/entities/descritor.entity';
@@ -9,20 +9,23 @@ export class CreateFonteDto {
   @ApiModelPropertyOptional({
     description: 'Nome da fonte',
   })
+  @IsOptional()
   @IsString()
-  nome: string;
+  nome?: string;
 
   @ApiModelPropertyOptional({
     description: 'URL da fonte',
   })
+  @IsOptional()
   @IsString()
-  link: string;
+  link?: string;
 
   @ApiModelPropertyOptional({
     description: 'Descrição da fonte',
   })
+  @IsOptional()
   @IsString()
-  descricao: string;
+  descricao?: string;
 
   @ApiModelProperty({
     description: 'Tipo da fonte',
@@ -35,13 +38,15 @@ export class CreateFonteDto {
     isArray: true,
     type: Assunto,
   })
-  assuntos: Assunto[];
+  @IsOptional()
+  assuntos?: Assunto[];
 
   @ApiModelPropertyOptional({
     description: 'Descritores relacionados a fonte',
     isArray: true,
     type: Descritor,
   })
-  descritores: Descritor[];
+  @IsOptional()
+  descritores?: Descritor[];
 
 }
