+++ src/modules/grupoAcesso/dto/grupoAcesso.dto.ts
@@ -1,4 +1,7 @@
 import { ApiModelProperty } from '@nestjs/swagger';
+import { NoticiaDto } from '../../noticiax/dto/index';
+import { Noticia } from '../../../database/entities/noticia.entity';
+import { UsuarioDto } from '../../usuario/dto/index';
 
 export class GrupoAcessoDto {
   @ApiModelProperty({
@@ -9,5 +12,19 @@ export class GrupoAcessoDto {
   @ApiModelProperty({
     description: 'Descrição do grupo de acesso',
   })
-  nome: string ;
+  nome: string;
+
+  @ApiModelProperty({
+    description: 'Noticias relacionadas ao grupo de acesso',
+    type: NoticiaDto,
+    isArray: true,
+  })
+  noticias?: NoticiaDto[];
+
+  @ApiModelProperty({
+    description: 'Usuarios do grupo de acesso',
+    type: UsuarioDto,
+    isArray: true,
+  })
+  usuarios?: UsuarioDto[];
 }
