+++ src/modules/midia/dto/updateMidia.dto.ts
@@ -1,5 +1,5 @@
 import { ApiModelPropertyOptional, ApiModelProperty } from '@nestjs/swagger';
-import { IsString, IsIn, IsInt, IsDefined } from 'class-validator';
+import { IsString, IsIn, IsInt, IsDefined, IsOptional } from 'class-validator';
 import { NoticiaDto } from '../../noticiax/dto/index';
 import { TipoMidiaDto } from '../../tipoMidia/dto/tipoMidia.dto';
 
@@ -14,6 +14,7 @@ export class UpdateMidiaDto {
   @ApiModelPropertyOptional({
     description: 'Nome da mídia',
   })
+  @IsOptional()
   @IsString()
   nome: string;
 
@@ -21,20 +22,20 @@ export class UpdateMidiaDto {
     description: 'URL da mídia',
   })
   @IsString()
-  @IsDefined()
+  @IsOptional()
   link: string;
 
   @ApiModelPropertyOptional({
     description: 'Noticia relacionada a mídia',
     type: NoticiaDto,
   })
-  @IsDefined()
+  @IsOptional()
   noticia: NoticiaDto;
 
   @ApiModelPropertyOptional({
     description: 'Tipo da mídia',
     type: TipoMidiaDto,
   })
-  @IsDefined()
+  @IsOptional()
   tipoMidia: TipoMidiaDto;
 }
