+++ src/noticia/dto/noticia.dto.ts
@@ -1,47 +1,47 @@
-import { ApiModelProperty } from '@nestjs/swagger';
+// import { ApiModelProperty } from '@nestjs/swagger';
 
-export class grupoAcessoDto {
-    @ApiModelProperty()
-    readonly nome?: string;
-  }
+// export class grupoAcessoDto {
+//     @ApiModelProperty()
+//     readonly nome?: string;
+//   }
 
-export class tipoFonteDto {
-    @ApiModelProperty()
-    readonly nome?: string;
-  }
+// export class tipoFonteDto {
+//     @ApiModelProperty()
+//     readonly nome?: string;
+//   }
 
-export class FonteDto {
-    @ApiModelProperty()
-    readonly link: string;
-    @ApiModelProperty()
-    readonly nome?: string;
-    @ApiModelProperty()
-    readonly descricao?: string;
-    @ApiModelProperty()
-    readonly tipoFonte?: tipoFonteDto;
-  }
+// export class FonteDto {
+//     @ApiModelProperty()
+//     readonly link: string;
+//     @ApiModelProperty()
+//     readonly nome?: string;
+//     @ApiModelProperty()
+//     readonly descricao?: string;
+//     @ApiModelProperty()
+//     readonly tipoFonte?: tipoFonteDto;
+//   }
 
-export class NoticiaDto {
-  @ApiModelProperty()
-  readonly link: string;
-  @ApiModelProperty()
-  readonly descricao?: string;
-  @ApiModelProperty()
-  readonly dataPublicacao?: string;
-  @ApiModelProperty()
-  readonly titulo?: [string];
-  @ApiModelProperty()
-  readonly conteudo?: string;
-  @ApiModelProperty()
-  readonly dataCriacao?: string;
-  @ApiModelProperty()
-  readonly dataAtualizacao?: string;
-  @ApiModelProperty()
-  readonly fonte?: FonteDto;
-  @ApiModelProperty()
-  readonly midias?: [];
-  @ApiModelProperty()
-  readonly grupoAcesso?: grupoAcessoDto;
-  @ApiModelProperty()
-  readonly descritores?: [];
-}
+// export class NoticiaDto {
+//   @ApiModelProperty()
+//   readonly link: string;
+//   @ApiModelProperty()
+//   readonly descricao?: string;
+//   @ApiModelProperty()
+//   readonly dataPublicacao?: string;
+//   @ApiModelProperty()
+//   readonly titulo?: [string];
+//   @ApiModelProperty()
+//   readonly conteudo?: string;
+//   @ApiModelProperty()
+//   readonly dataCriacao?: string;
+//   @ApiModelProperty()
+//   readonly dataAtualizacao?: string;
+//   @ApiModelProperty()
+//   readonly fonte?: FonteDto;
+//   @ApiModelProperty()
+//   readonly midias?: [];
+//   @ApiModelProperty()
+//   readonly grupoAcesso?: grupoAcessoDto;
+//   @ApiModelProperty()
+//   readonly descritores?: [];
+// }
